Updated file
Added map functionality to show where the player is, as well as when there is a locked gate. Also added multiple stories and the ability to use stairs (but not the ability to tell whether those stairs are going up or down). Moved a significant amount of variables into a json array. Currently everything is still in the project in comment form, but eventually this should let me clean up the code and make it easier to deal with.